Now use tilesets to generate the map. Will allow me to quickly paint large detailed playing canvases. I want to try and some of the advanced tools such as setting up rules which will automatically paint the correct edge tiles.
had some issues with the per pixel size I had set for my sprites when setting this against the grid for the tileset. Tile sprites are now correct, but it broke the sizing of the characters and zombies. This has been somewhat fixed now.

Will try adding in a new gun next time.